+++ src/hks_pylib/cryptography/batchcrypt/batchnumber.py
@@ -1,39 +1,51 @@
 from hks_pylib.cryptography.batchcrypt.integer import SignedInteger
 from hks_pylib.errors import InvalidParameterError
+from hks_pylib.errors.cryptography.batchcrypt.integer import OverflowIntegerError
+from hks_pylib.errors.cryptography.batchcrypt.quantization import OverflowQuantizerError
 from hks_pylib.math import Bitwise
-from hks_pylib.cryptography.batchcrypt.quantization import GenericQuantizer, SignedQuantizer
+from hks_pylib.cryptography.batchcrypt.quantization import Quantizer
 
 from hks_pylib.errors.cryptography.batchcrypt.batchnumber import *
 
 class BatchNumber(object):
-    PADDING_SIZE = 2
+    __PADDING_SIZE = 2
+
+    def padding_size():
+        return BatchNumber.__PADDING_SIZE
+    
+    def set_padding_size(value: int):
+        if not isinstance(value, int) or value < 0:
+            raise InvalidParameterError("Padding size must be a "
+            "positive int.")
+
+        BatchNumber.__PADDING_SIZE = value
 
     def __init__(
                 self,
                 batch: int,
-                num: int,   # the number of elements in the batch
-                size: int   # the size of each element in the batch
+                num: int,   # the number of values in the batch
+                size: int   # the size of each values in the batch
             ) -> None:
         self._num = num
-        self._size = size
-        self._total_size = size + BatchNumber.PADDING_SIZE
+        self._value_size = size
+        self._element_size = size + BatchNumber.padding_size()
 
-        self._batch = batch
-        self._batch_size = self._num * self._total_size
+        self._raw = batch
+        self._total_size = self._num * self._element_size
 
     def get(self, index: int) -> int:
         if not isinstance(index, int):
             raise InvalidElementBatchNumberError("Parameter index must be an int.")
 
-        if index >= self._num or index < 0:
+        if index >= len(self) or index < 0:
             raise InvalidElementBatchNumberError("You can only access elements "
-            "starting from 0 to {}".format(self._num - 1))
+            "starting from 0 to {}".format(len(self) - 1))
 
         return Bitwise.get_bits(
-                number=self._batch,
-                position=self._batch_size\
-                    - index * self._total_size - 1,
-                length=self._total_size
+                number=self.raw(),
+                position=self.total_size()\
+                    - index * self.element_size() - 1,
+                length=self.element_size()
             )
 
     def __getitem__(self, index: int):
@@ -42,172 +54,198 @@ class BatchNumber(object):
     def __len__(self):
         return self._num
 
-    def size(self):
-        return self._size
+    def value_size(self):
+        "Only the size of value without padding bits."
+        return self._value_size
+
+    def element_size(self):
+        "Including both value size and padding size."
+        return self._element_size
 
     def total_size(self):
         return self._total_size
 
+    def raw(self):
+        return self._raw
+
     def __add__(self, other):
         if not isinstance(other, type(self)):
             return InvalidParameterError("Two operands "
-            "must be the same type ({} != {}).".format(type(self), type(other)))
+            "must be the same type ({} != {}).".format(
+                type(self).__name__, type(other).__name__))
 
-        if self._num != other._num:
+        if len(self) != len(other):
             raise MismatchedTypeBatchNumberError("Two operands "
             "must be the same number of elements ({} != {}).".format(
-                    self._num,
-                    other._num
+                    len(self),
+                    len(other)
                 ))
 
-        if self._total_size != other._total_size:
+        if self.element_size() != other.element_size():
             raise MismatchedTypeBatchNumberError("Two operands "
             "must be the same element length ({} != {}).".format(
-                    self._total_size,
-                    other._total_size
+                    self.element_size(),
+                    other.element_size()
                 ))
 
         return BatchNumber(
-                self._batch + other._batch,
-                self._num,
-                self._size
+                self.raw() + other.raw(),
+                len(self),
+                self.value_size()
             )
 
     
     def __iadd__(self, other):
         if not isinstance(other, type(self)):
             return InvalidParameterError("Two operands "
-            "must be the same type ({} != {}).".format(type(self), type(other)))
+            "must be the same type ({} != {}).".format(
+                type(self).__name__, type(other).__name__))
 
-        if self._num != other._num:
+        if len(self) != len(other):
             raise MismatchedTypeBatchNumberError("Two operands "
             "must be the same number of elements ({} != {}).".format(
-                    self._num,
-                    other._num
+                    len(self),
+                    len(other)
                 ))
 
-        if self._total_size != other._total_size:
+        if self.element_size() != other.element_size():
             raise MismatchedTypeBatchNumberError("Two operands "
             "must be the same element length ({} != {}).".format(
-                    self._total_size,
-                    other._total_size
+                    self.element_size(),
+                    other.element_size()
                 ))
 
-        self._batch += other._batch
+        self._raw += other.raw()
+
+    def __len__(self):
+        return self._num
 
     def __str__(self) -> str:
-        return bin(self._batch)
+        return bin(self._raw)
 
     def __repr__(self) -> str:
         return str(self)
-    
-    def to_int(self) -> int:
-        return self._batch
 
     @staticmethod
-    def bind(*numbers, size: int = 8):
-        total_size = size + BatchNumber.PADDING_SIZE
+    def bind(*numbers, size: int):
+        element_size = size + BatchNumber.padding_size()
 
-        batch_length = total_size * len(numbers)
-        batch = 0
+        total_size = element_size * len(numbers)
+        raw = 0
 
         for i, number in enumerate(numbers):
-            batch = Bitwise.set_bits(
-                    number=batch,
-                    position=batch_length - total_size * i\
-                        - BatchNumber.PADDING_SIZE - 1,
+            raw = Bitwise.set_bits(
+                    number=raw,
+                    position=total_size - element_size * i\
+                        - BatchNumber.padding_size() - 1,
                     length=size,
                     value=number
                 )
 
-        return BatchNumber(batch, len(numbers), size)
-
+        return BatchNumber(raw, len(numbers), size)
 
 
 class GenericBatchNumber(BatchNumber):
-    __quantizer = GenericQuantizer()
+    __quantizer = Quantizer()
     def __init__(self, batch: int, num: int, size: int) -> None:
         super().__init__(batch, num, size)
         self._n_cumulative = 1
     
     @staticmethod
-    def quantizer():
-        return GenericBatchNumber.__quantizer
-
-    def __add__(self, other: BatchNumber):
-        if not isinstance(other, GenericBatchNumber):
-            raise InvalidParameterError("Operands of addition "
-            "must be GenericBatchNumber objects.")
-
-        result = super().__add__(other)
-        result = GenericBatchNumber(
-                result._batch,
-                result._num,
-                result._size
-            )
-        result._n_cumulative = self._n_cumulative + other._n_cumulative
-        return result
-    
-    def __iadd__(self, other: BatchNumber):
-        if not isinstance(other, GenericBatchNumber):
-            raise InvalidParameterError("Operands of addition "
-            "must be GenericBatchNumber objects.")
-
-        super().__iadd__(other)
-        self._n_cumulative += other._n_cumulative
-
-    def get(self, index: int) -> int:
-        result = super().get(index)
-        return self.__quantizer.i2f(result, n_cumulative=self._n_cumulative)
+    def quantizer(float_range: tuple, int_size: int):
+        GenericBatchNumber.__quantizer.set_float_range(*float_range)
+        GenericBatchNumber.__quantizer.set_int_size(int_size, signed=False)
+        GenericBatchNumber.__quantizer.compile()
 
     @staticmethod
     def bind(*numbers, size: int):
         numbers = (GenericBatchNumber.__quantizer.f2i(f) for f in numbers)
         result = BatchNumber.bind(*numbers, size=size)
-        return GenericBatchNumber(result._batch, result._num, result._size)
+        return GenericBatchNumber(result._raw, result._num, result.value_size())
+
+    def get(self, index: int) -> int:
+        result = super().get(index)
+        return self.__quantizer.i2f(result, n_cumulative=self._n_cumulative)
+
+    def __add__(self, other: BatchNumber):
+        if not isinstance(other, type(self)):
+            raise InvalidParameterError("Operands of addition "
+            "must be {} objects.".format(type(self).__name__))
+
+        result = super().__add__(other)
+        result = GenericBatchNumber(
+                result.raw(),
+                len(result),
+                result.value_size()
+            )
+        result._n_cumulative = self._n_cumulative + other._n_cumulative
+        return result
+
+    def __iadd__(self, other: BatchNumber):
+        if not isinstance(other, type(self)):
+            raise InvalidParameterError("Operands of addition "
+            "must be {} objects.".format(type(self).__name__))
+
+        super().__iadd__(other)
+        self._n_cumulative += other._n_cumulative
 
 
 class SignedBatchNumber(BatchNumber):
-    __quantizer = SignedQuantizer()
+    __quantizer = Quantizer()
 
     def __init__(self, batch: int, num: int, size: int) -> None:
         super().__init__(batch, num, size)
-        self._original_size = size - SignedInteger.sign_size() + 1
 
     @staticmethod
-    def quantizer():
-        return SignedBatchNumber.__quantizer
+    def quantizer(float_range: tuple, int_size: int):
+        if not isinstance(float_range, tuple) or len(float_range) != 2\
+            or sum(float_range) != 0:
+            raise InvalidParameterError("Float range "
+            "in signed batch number must be symmetric")
 
-    @staticmethod
-    def bind(*numbers, size: int):
-        signed_numbers = []
-        for n in numbers:
-            sn = SignedBatchNumber.__quantizer.f2i(n)
-            signed_numbers.append(sn.to_int())
-
-        raw_result = BatchNumber.bind(
-                *signed_numbers, 
-                size=SignedInteger.total_size_(size)
+        SignedBatchNumber.__quantizer.set_float_range(*float_range)
+        SignedBatchNumber.__quantizer.set_int_size(
+                int_size - SignedInteger.sign_size() + 1,
+                signed=True
             )
+        SignedBatchNumber.__quantizer.compile()
+
+    @staticmethod
+    def bind(*fnumbers, size: int):
+        signed_number_list = []
+        for f in fnumbers:
+            i = SignedBatchNumber.__quantizer.f2i(f)
+            signed_number: SignedInteger = SignedInteger.from_int(i, size)
+            signed_number_list.append(signed_number.raw())
+
+        super_result = BatchNumber.bind(
+                *signed_number_list, 
+                size=size
+            )
+
         result = SignedBatchNumber(
-                batch=raw_result._batch,
-                num=raw_result._num,
-                size=raw_result._size
+                batch=super_result.raw(),
+                num=len(super_result),
+                size=super_result.value_size()
             )
         return result
 
     def get(self, index: int) -> int:
         raw_result = super().get(index)
-        signed_number = SignedInteger(
-                number=raw_result,
-                original_size=self._original_size
-            )
-        return self.__quantizer.i2f(signed_number)
+        try:
+            signed_number = SignedInteger(
+                    number=raw_result,
+                    size=self.value_size()
+                )
+        except OverflowIntegerError as e:
+            raise e
+
+        return self.__quantizer.i2f(signed_number.value())
 
     def __add__(self, other: BatchNumber):
         raw_result = super().__add__(other)
         return SignedBatchNumber(
-                batch=raw_result._batch,
-                num=raw_result._num,
-                size=raw_result._size
+                batch=raw_result.raw(),
+                num=len(raw_result),
+                size=raw_result.value_size()
             )